Fix swiftlint warnings
