+++ PerformanceSuite.podspec
@@ -24,7 +24,7 @@ Pod::Spec.new do |s|
 
   # Sample App which is also used for UI tests as a host
   s.app_spec 'PerformanceApp' do |app_spec|
-    app_spec.source_files = 'PerformanceSuite/PerformanceApp/*.swift'
+    app_spec.source_files = 'PerformanceSuite/PerformanceApp/**/*.swift'
     app_spec.dependency 'GCDWebServer'
     app_spec.dependency 'PerformanceSuite/Crashlytics'
   end
@@ -32,7 +32,7 @@ Pod::Spec.new do |s|
   # Unit tests
   s.test_spec 'Tests' do |test_spec|
     test_spec.requires_app_host = true
-    test_spec.source_files = 'PerformanceSuite/Tests/*.swift'
+    test_spec.source_files = 'PerformanceSuite/Tests/**/*.swift'
     test_spec.dependency 'PerformanceSuite/Crashlytics'
   end
 
@@ -40,7 +40,7 @@ Pod::Spec.new do |s|
   s.test_spec 'UITests' do |test_spec|
     test_spec.requires_app_host = true
     test_spec.test_type = :ui
-    test_spec.source_files = 'PerformanceSuite/UITests/*.swift', 'PerformanceSuite/PerformanceApp/UITestsInterop.swift'
+    test_spec.source_files = 'PerformanceSuite/UITests/**/*.swift', 'PerformanceSuite/PerformanceApp/UITestsInterop.swift'
     test_spec.app_host_name = 'PerformanceSuite/PerformanceApp'
     test_spec.dependency 'PerformanceSuite/PerformanceApp'
     test_spec.dependency 'GCDWebServer'
