+++ PerformanceSuite/Tests/.swiftlint.yml
@@ -1,0 +1,10 @@
+disabled_rules:
+  - force_unwrapping
+
+type_body_length:
+  warning: 500
+  error: 500
+
+file_length:
+  warning: 700
+  error: 700
