+++ PerformanceSuite/Tests/CrashlyticsIssueReporterTests.swift
@@ -3,9 +3,8 @@ import XCTest
 @testable import PerformanceSuite
 
 final class CrashlyticsIssueReporterTests: XCTestCase {
-    func testStackTraceParsing() {
-        let reporter = CrashlyticsIssueReporter(fatalHangsAsCrashes: true, firebaseHangReason: "reason")
-        let stack = """
+
+    let stack = """
             0   libsystem_kernel.dylib         0x00000002036a2c88 0x00000002036a1000 + 8
             1   libsystem_pthread.dylib        0x0000000224325114 0x0000000224323000 + 84
             2   libsystem_pthread.dylib        0x000000022432c318 0x0000000224323000 + 248
@@ -65,65 +64,68 @@ final class CrashlyticsIssueReporterTests: XCTestCase {
             56  dyld                           0x00000001e3e9cdec 0x00000001e3e87000 + 2220
             """
 
-        let addresses: [UInt] = [
-            0x0000_0002_036a_2c88,
-            0x0000_0002_2432_5114,
-            0x0000_0002_2432_c318,
-            0x0000_0001_c822_c7e4,
-            0x0000_0001_c82a_2a00,
-            0x0000_0001_c822_aec0,
-            0x0000_0001_c6a8_4898,
-            0x0000_0001_c6c1_1508,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_1564,
-            0x0000_0001_c6c1_14d8,
-            0x0000_0001_c6c1_38dc,
-            0x0000_0001_c6c1_1344,
-            0x0000_0001_c6c1_1130,
-            0x0000_0001_c6ac_6e54,
-            0x0000_0001_036e_0d4c,
-            0x0000_0001_037b_7518,
-            0x0000_0001_00fb_e7c0,
-            0x0000_0001_cbe5_1eac,
-            0x0000_0001_cbe5_5330,
-            0x0000_0001_cbe6_8908,
-            0x0000_0001_cbe6_05f8,
-            0x0000_0001_cbe6_02f4,
-            0x0000_0001_c4a2_1d18,
-            0x0000_0001_c4a0_3650,
-            0x0000_0001_c4a0_84dc,
-            0x0000_0001_ffc7_435c,
-            0x0000_0001_c6d9_437c,
-            0x0000_0001_c6d9_3fe0,
-            0x0000_0001_00f4_9738,
-            0x0000_0001_e3e9_cdec,
-        ]
+    let addresses: [UInt] = [
+        0x0000_0002_036a_2c88,
+        0x0000_0002_2432_5114,
+        0x0000_0002_2432_c318,
+        0x0000_0001_c822_c7e4,
+        0x0000_0001_c82a_2a00,
+        0x0000_0001_c822_aec0,
+        0x0000_0001_c6a8_4898,
+        0x0000_0001_c6c1_1508,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_1564,
+        0x0000_0001_c6c1_14d8,
+        0x0000_0001_c6c1_38dc,
+        0x0000_0001_c6c1_1344,
+        0x0000_0001_c6c1_1130,
+        0x0000_0001_c6ac_6e54,
+        0x0000_0001_036e_0d4c,
+        0x0000_0001_037b_7518,
+        0x0000_0001_00fb_e7c0,
+        0x0000_0001_cbe5_1eac,
+        0x0000_0001_cbe5_5330,
+        0x0000_0001_cbe6_8908,
+        0x0000_0001_cbe6_05f8,
+        0x0000_0001_cbe6_02f4,
+        0x0000_0001_c4a2_1d18,
+        0x0000_0001_c4a0_3650,
+        0x0000_0001_c4a0_84dc,
+        0x0000_0001_ffc7_435c,
+        0x0000_0001_c6d9_437c,
+        0x0000_0001_c6d9_3fe0,
+        0x0000_0001_00f4_9738,
+        0x0000_0001_e3e9_cdec,
+    ]
+
+    func testStackTraceParsing() {
+        let reporter = CrashlyticsIssueReporter(fatalHangsAsCrashes: true, firebaseHangReason: "reason")
 
         let expectedFrames = addresses.map { StackFrame(address: $0) }
         let frames = reporter.parse(stackTrace: stack)
