+++ PerformanceSuite/Tests/HangReporterTests.swift
@@ -9,9 +9,6 @@ import XCTest
 
 @testable import PerformanceSuite
 
-// swiftlint:disable force_unwrapping
-// swiftlint:disable file_length
-// swiftlint:disable type_body_length
 class HangReporterTests: XCTestCase {
 
     override func setUp() {
