+++ PerformanceSuite/Tests/TTIObserverExtensionTests.swift
@@ -44,7 +44,7 @@ class TTIObserverExtensionTests: XCTestCase {
         window.makeKeyAndVisible()
 
         let exp = expectation(description: "vc appeared")
-        let vc = MyViewController()
+        let vc = OutputViewController()
         vc.viewAppeared = {
             exp.fulfill()
         }
@@ -74,11 +74,11 @@ class TTIObserverExtensionTests: XCTestCase {
     }
 
     func testTTIIsFinishedWhenViewDisappearedByDefault() {
-        let root = MyViewController()
+        let root = OutputViewController()
         root.title = "root"
-        let vc1 = MyViewController()
+        let vc1 = OutputViewController()
         vc1.title = "vc1"
-        let vc2 = MyViewController()
+        let vc2 = OutputViewController()
         vc2.title = "vc2"
         let navigation = UINavigationController(rootViewController: root)
 
@@ -133,7 +133,7 @@ class TTIObserverExtensionTests: XCTestCase {
     }
 
     func testNoTTIForViewControllerWithoutScreenIsReady() {
-        let vc = MyViewController()
+        let vc = OutputViewController()
         vc.title = "vc"
 
         let exp = expectation(description: "viewDidAppear")
@@ -163,7 +163,7 @@ class TTIObserverExtensionTests: XCTestCase {
     }
 
     func testScreenIsReadyForViewControllerGeneratesTTIMetrics() {
-        let vc = MyViewController()
+        let vc = OutputViewController()
         vc.title = "vc"
 
         let exp = expectation(description: "viewDidAppear")
@@ -200,7 +200,7 @@ class TTIObserverExtensionTests: XCTestCase {
     }
 
     func testScreenIsReadyForSwiftUIViewGeneratesTTIMetrics() {
-        let vc = MyHostingController(rootView: MyView())
+        let vc = HostingControllerWithAppeared(rootView: ViewIsReadyOnAppear())
         vc.title = "hosting vc"
 
         let window = makeWindow()
@@ -225,13 +225,13 @@ class TTIObserverExtensionTests: XCTestCase {
         let now = makeRandomTime()
         timeProvider.time = now
 
-        let vc1 = MyViewController()
+        let vc1 = OutputViewController()
         vc1.title = "vc1"
 
-        let vc2 = MyViewController()
+        let vc2 = OutputViewController()
         vc2.title = "vc2"
 
-        let vc3 = MyViewController()
+        let vc3 = OutputViewController()
         vc3.title = "vc3"
 
         PerformanceMonitoring.queue.sync {}
@@ -329,10 +329,10 @@ class TTIObserverExtensionTests: XCTestCase {
         let now = makeRandomTime()
         timeProvider.time = now
 
-        let vc1 = MyViewController()
+        let vc1 = OutputViewController()
         vc1.title = "vc1"
 
-        let vc2 = UIViewController()
+        let vc2 = OutputViewController()
         vc2.title = "vc2"
 
         vc1.addChild(vc2)
@@ -371,72 +371,5 @@ class TTIObserverExtensionTests: XCTestCase {
 
     private func makeRandomTime() -> DispatchTime {
         DispatchTime(uptimeNanoseconds: 10_000_000 + UInt64.random(in: 0..<100000))
-    }
-}
-
-private class MyHostingController<T: View>: UIHostingController<T> {
-
-    var viewAppeared: () -> Void = {}
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        DispatchQueue.main.async {
-            self.viewAppeared()
-        }
-    }
-}
-
-private class MyViewController: UIViewController {
-
-    var viewDisappeared: () -> Void = {}
-    var viewAppeared: () -> Void = {}
-    var output = ""
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        output += "viewDidLoad\n"
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        output += "viewWillAppear\n"
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        output += "viewDidAppear\n"
-        viewAppeared()
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        output += "viewWillDisappear\n"
-    }
-
-    override func viewDidDisappear(_ animated: Bool) {
-        super.viewDidDisappear(animated)
-        output += "viewDidDisappear\n"
-
-        viewDisappeared()
-    }
-
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        output += "viewWillLayoutSubviews\n"
-    }
-
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        output += "viewDidLayoutSubviews\n"
-    }
-
-    deinit {
-        XCTAssertTrue(Thread.isMainThread)
-    }
-}
-
-private struct MyView: View {
-    var body: some View {
-        return Text("test").screenIsReadyOnAppear()
     }
 }
