+++ PerformanceSuite/Tests/Utils/ViewIsReadyOnAppear.swift
@@ -1,0 +1,14 @@
+//
+//  ViewIsReadyOnAppear.swift
+//  Pods
+//
+//  Created by Gleb Tarasov on 21/09/2024.
+//
+
+import SwiftUI
+
+struct ViewIsReadyOnAppear: View {
+    var body: some View {
+        return Text("test").screenIsReadyOnAppear()
+    }
+}
